mysite/requirements.txt
# Dependencies to use in container; always give it a fixed version!
Django==3.2
psycopg2-binary==2.9.3
djangorestframework==3.13.1
Pygments==2.12.0
daphne==3.0.2
channels==3.0.4
channels-redis==3.4.0
